fixup sequence container and put bps in columns
